% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRatio.R
\name{getRatio}
\alias{getRatio}
\title{Divide two columns and create a new column}
\usage{
getRatio(df, i, j, k)
}
\arguments{
\item{df}{A data frame.}

\item{i}{A column of a data frame.}

\item{j}{A column of a data frame.}

\item{k}{A new column of a data frame with a specified name.}
}
\value{
A data frame containing a new column k, containing ratio data from dividing column i by column j.
}
\description{
This function divides each row of a specified column by each row of another specified column. Both columns must be from the same data frame and the first argument must be a data frame.
Any row with NA will be dropped. The function will output a data frame containing a new column with the desired results.
}
